--- a/binding.gyp
+++ b/binding.gyp
@@ -8,13 +8,13 @@
       ],
       'defines': [ 'NAPI_DISABLE_CPP_EXCEPTIONS' ],
       "cflags": [ "-fno-exceptions" ],
-      "cflags_cc": [ "-fno-exceptions", "-std=c++17" ],
+      "cflags_cc": [ "-fno-exceptions", "-std=c++2a" ],
       "conditions": [
         ['OS=="mac"', {
             "xcode_settings": {
               'GCC_ENABLE_CPP_EXCEPTIONS': 'YES',
               "CLANG_CXX_LIBRARY": "libc++",
-              "CLANG_CXX_LANGUAGE_STANDARD":"c++17",
+              "CLANG_CXX_LANGUAGE_STANDARD":"c++2a",
               'MACOSX_DEPLOYMENT_TARGET': '10.15'
             }
         }],