--- a/src/common.h
+++ b/src/common.h
@@ -95,7 +95,7 @@ extern Score path_scorer_score(const CandidateString &string, const Element &que
 extern int countDir(const CandidateString &path, int end, char pathSeparator);
 extern CandidateString getExtension(const CandidateString &str);
 
-extern std::vector<CandidateIndex> filter(const vector<std::vector<CandidateString>> &candidates, const Element &query, const Options &options);
+extern const std::vector<CandidateIndex> filter(const vector<std::vector<CandidateString>> &candidates, const Element &query, const Options &options);
 
 extern std::vector<size_t> matcher_match(const CandidateString &string, const Element &query, const Options &options);
 extern void get_wrap(const CandidateString &string, const Element &query, const Options &options, std::string *out);