--- a/src/fuzzaldrin.cc
+++ b/src/fuzzaldrin.cc
@@ -3,7 +3,7 @@
 #include "fuzzaldrin.h"
 
 Napi::Value Fuzzaldrin::Filter(const Napi::CallbackInfo &info) {
-    Napi::Array res = Napi::Array::New(info.Env());
+    auto res = Napi::Array::New(info.Env());
     if (info.Length() != 4 || !info[0].IsString() || !info[1].IsNumber() || !info[2].IsBoolean() || !info[3].IsBoolean()) {
         Napi::TypeError::New(info.Env(), "Invalid arguments").ThrowAsJavaScriptException();
         return Napi::Boolean();
@@ -26,19 +26,19 @@ Napi::Value Fuzzaldrin::SetCandidates(const Napi::CallbackInfo &info) {
         Napi::TypeError::New(info.Env(), "Invalid arguments").ThrowAsJavaScriptException();
         return Napi::Boolean();
     }
-    Napi::Array candidates = info[0].As<Napi::Array>();
+    auto candidates = info[0].As<Napi::Array>();
     const size_t N = candidates.Length();
-    const size_t num_chunks = (N < 1000 * kMaxThreads) ? (N / 1000 + 1) : kMaxThreads;
+    const auto num_chunks = N < 1000 * kMaxThreads ? N / 1000 + 1 : kMaxThreads;
     candidates_.clear();
     candidates_.resize(num_chunks);
     size_t cur_start = 0;
     for (size_t i = 0; i < num_chunks; i++) {
-        size_t chunk_size = N / num_chunks;
+        auto chunk_size = N / num_chunks;
         // Distribute remainder among the chunks.
         if (i < N % num_chunks) {
             chunk_size++;
         }
-        for (size_t j = cur_start; j < cur_start + chunk_size; j++) {
+        for (auto j = cur_start; j < cur_start + chunk_size; j++) {
             Napi::Value val = candidates[j];
             candidates_[i].push_back(val.ToString().Utf8Value());
         }
@@ -48,18 +48,18 @@ Napi::Value Fuzzaldrin::SetCandidates(const Napi::CallbackInfo &info) {
 }
 
 void Fuzzaldrin::SetCandidates(vector<CandidateObject> const &candidates) {
-    const size_t N = candidates.size();// different
-    const size_t num_chunks = (N < 1000 * kMaxThreads) ? (N / 1000 + 1) : kMaxThreads;
+    const auto N = candidates.size();// different
+    const auto num_chunks = N < 1000 * kMaxThreads ? N / 1000 + 1 : kMaxThreads;
     candidates_.clear();
     candidates_.resize(num_chunks);
     size_t cur_start = 0;
     for (size_t i = 0; i < num_chunks; i++) {
-        size_t chunk_size = N / num_chunks;
+        auto chunk_size = N / num_chunks;
         // Distribute remainder among the chunks.
         if (i < N % num_chunks) {
             chunk_size++;
         }
-        for (size_t j = cur_start; j < cur_start + chunk_size; j++) {
+        for (auto j = cur_start; j < cur_start + chunk_size; j++) {
             candidates_[i].push_back(candidates[j].data);// different
         }
         cur_start += chunk_size;
@@ -96,7 +96,7 @@ Napi::Value Fuzzaldrin::FilterTree(const Napi::CallbackInfo &info) {
     const auto matches = filter(candidates_, query, options);
 
     // filter
-    Napi::Array filteredCandidateObjects = Napi::Array::New(info.Env());// array of candidate objects (with their address in index and level)
+    auto filteredCandidateObjects = Napi::Array::New(info.Env());// array of candidate objects (with their address in index and level)
     for (uint32_t i = 0, len = matches.size(); i < len; i++) {
         auto entry = tree.entriesArray[matches[i]];//
 
@@ -127,7 +127,7 @@ Napi::Number score(const Napi::CallbackInfo &info) {
 }
 
 Napi::Array match(const Napi::CallbackInfo &info) {
-    Napi::Array res = Napi::Array::New(info.Env());
+    auto res = Napi::Array::New(info.Env());
     if (info.Length() != 3 || !info[0].IsString() || !info[1].IsString() || !info[2].IsString()) {
         Napi::TypeError::New(info.Env(), "Invalid arguments").ThrowAsJavaScriptException();
         return res;
@@ -168,7 +168,7 @@ Napi::String wrap(const Napi::CallbackInfo &info) {
 Napi::Object Fuzzaldrin::Init(Napi::Env env, Napi::Object exports) {
     Napi::HandleScope scope(env);
 
-    Napi::Function func = DefineClass(env, "Fuzzaldrin", { InstanceMethod("filter", &Fuzzaldrin::Filter), InstanceMethod("filterTree", &Fuzzaldrin::FilterTree), InstanceMethod("setCandidates", &Fuzzaldrin::SetCandidates) });
+    const auto func = DefineClass(env, "Fuzzaldrin", { InstanceMethod("filter", &Fuzzaldrin::Filter), InstanceMethod("filterTree", &Fuzzaldrin::FilterTree), InstanceMethod("setCandidates", &Fuzzaldrin::SetCandidates) });
 
     exports.Set("Fuzzaldrin", func);
     exports.Set("score", Napi::Function::New(env, score));