- Add and run clang-tidy
- Add auto where possible
- Add const where possible
- Build on C++20 on all platforms
- Test on old platforms

<details>
<summary>Benchmarks before and after</summary>

TwoLetter#Keybased#Filter seems to be faster now. In other benchmarks, there is no meaningful performance difference. Multiple benchmarks give different results.

Benchmarks after:

```
> fuzzaldrin-plus-fast@1.1.0 benchmark C:\Users\aminy\Documents\GitHub\JavaScript\@atom-ide-community\fuzzaldrin-plus-fast
> npm run benchmark:small && npm run benchmark:regular && npm run benchmark:large


> fuzzaldrin-plus-fast@1.1.0 benchmark:small
> node benchmark/benchmark-small.js

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 100 100

====== Running test - query:npm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 4 ms
length 55 100

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 100 100

====== Running test - query:grunt ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 33 100

====== Running test - query:html ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 10 100

====== Running test - query:doc ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 4 ms
length 87 100

====== Running test - query:cli ======
Elapsed time - fuzzaldrin-plus-fast: 1 ms vs. fuzzaldrin-plus: 2 ms
length 57 100

====== Running test - query:js ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 60 100

====== Running test - query:jas ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 19 100

====== Running test - query:mine ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 65 100

====== Running test - query:stream ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 19 100


> fuzzaldrin-plus-fast@1.1.0 benchmark:regular
> node benchmark/benchmark.js

====== Running test - query:index ======
Elapsed time - fuzzaldrin-plus-fast: 30 ms vs. fuzzaldrin-plus: 44 ms
length 6168 66672

====== Running test - query:indx ======
Elapsed time - fuzzaldrin-plus-fast: 30 ms vs. fuzzaldrin-plus: 54 ms
length 6192 66672

====== Running test - query:walkdr ======
Elapsed time - fuzzaldrin-plus-fast: 26 ms vs. fuzzaldrin-plus: 16 ms
length 504 66672
====== fuzzaldrin-plus-fast is SLOWER

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 48 ms vs. fuzzaldrin-plus: 68 ms
length 65136 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 47 ms vs. fuzzaldrin-plus: 66 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 47 ms vs. fuzzaldrin-plus: 50 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 48 ms vs. fuzzaldrin-plus: 60 ms
length 65208 66672

====== Running test - query:ndem ======
Elapsed time - fuzzaldrin-plus-fast: 52 ms vs. fuzzaldrin-plus: 236 ms
length 65124 66672

Matching 66672 results for 'index' took 297ms (Prepare in advance)
Matching 66672 results for 'index' took 294ms (cache)
Matching 66672 results for 'index' took 84ms (legacy)

> fuzzaldrin-plus-fast@1.1.0 benchmark:large
> node benchmark/benchmark-large.js

TwoLetter#legacy: 11.107s
TwoLetter#fuzzaldrin-plus-fast#DirectFilter: 3.201s
TwoLetter#fuzzaldrin-plus-fast#setCandidates#filter: 478.81ms
======
ThreeLetter#legacy: 8.747s
ThreeLetter#fuzzaldrin-plus-fast#DirectFilter: 3.443s
ThreeLetter#fuzzaldrin-plus-fast#setCandidates#filter: 451.884ms
======
TwoLetter#Keybased#Filter: 4.180s
ThreeLetter#Keybased#Filter: 4.878s
======
setCandidates: 240.143ms
TwoLetter#Filter: 468.944ms
ThreeLetter#Filter: 451.256ms
======
setCandidates#Keybased: 246.632ms
TwoLetter#Keybased#Filter: 471.995ms
ThreeLetter#Keybased#Filter: 443.108ms
```
Benchmarks before:
```
> npm run benchmark:small && npm run benchmark:regular && npm run benchmark:large


> fuzzaldrin-plus-fast@1.1.0 benchmark:small
> node benchmark/benchmark-small.js

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 100 100

====== Running test - query:npm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 4 ms
length 55 100

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 100 100

====== Running test - query:grunt ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 33 100

====== Running test - query:html ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 10 100

====== Running test - query:doc ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 4 ms
length 87 100

====== Running test - query:cli ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 57 100

====== Running test - query:js ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 60 100

====== Running test - query:jas ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 19 100

====== Running test - query:mine ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 65 100

====== Running test - query:stream ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 19 100


> fuzzaldrin-plus-fast@1.1.0 benchmark:regular
> node benchmark/benchmark.js

====== Running test - query:index ======
Elapsed time - fuzzaldrin-plus-fast: 29 ms vs. fuzzaldrin-plus: 45 ms
length 6168 66672

====== Running test - query:indx ======
Elapsed time - fuzzaldrin-plus-fast: 29 ms vs. fuzzaldrin-plus: 51 ms
length 6192 66672

====== Running test - query:walkdr ======
Elapsed time - fuzzaldrin-plus-fast: 27 ms vs. fuzzaldrin-plus: 16 ms
length 504 66672
====== fuzzaldrin-plus-fast is SLOWER

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 50 ms vs. fuzzaldrin-plus: 68 ms
length 65136 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 51 ms vs. fuzzaldrin-plus: 60 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 49 ms vs. fuzzaldrin-plus: 52 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 48 ms vs. fuzzaldrin-plus: 54 ms
length 65208 66672

====== Running test - query:ndem ======
Elapsed time - fuzzaldrin-plus-fast: 51 ms vs. fuzzaldrin-plus: 242 ms
length 65124 66672

Matching 66672 results for 'index' took 298ms (Prepare in advance)
Matching 66672 results for 'index' took 291ms (cache)
Matching 66672 results for 'index' took 89ms (legacy)

> fuzzaldrin-plus-fast@1.1.0 benchmark:large
> node benchmark/benchmark-large.js

TwoLetter#legacy: 11.255s
TwoLetter#fuzzaldrin-plus-fast#DirectFilter: 3.268s
TwoLetter#fuzzaldrin-plus-fast#setCandidates#filter: 485.822ms
======
ThreeLetter#legacy: 8.887s
ThreeLetter#fuzzaldrin-plus-fast#DirectFilter: 3.281s
ThreeLetter#fuzzaldrin-plus-fast#setCandidates#filter: 446.293ms
======
TwoLetter#Keybased#Filter: 3.857s
ThreeLetter#Keybased#Filter: 4.677s
======
setCandidates: 235.303ms
TwoLetter#Filter: 646.962ms
ThreeLetter#Filter: 459.873ms
======
setCandidates#Keybased: 233.968ms
TwoLetter#Keybased#Filter: 638.149ms
ThreeLetter#Keybased#Filter: 461.798ms
```

</details>